packages/shared-backend/pyproject.toml
[tool.poetry]
name = "shared-backend"
version = "1.0.0"
description = "Shared backend utilities for KVSHVL platform"
authors = ["Kushal Samant"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
sqlalchemy = "^2.0.36"
pydantic = "^2.10.4"
pydantic-settings = "^2.6.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
razorpay = "^1.4.2"
python-dotenv = "^1.0.0"
psycopg = {extras = ["binary"], version = "^3.2.1"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

